Move code to get email to send confirmation to static function
This will enable us to use the same function to display which address the email has been sent to. This is called after a redirect, would be better if we had an email object we could pass around but I think this may be the best we can do here.

Signed-off-by: Mark Muhleder <mark@analogweb.co.uk>